Updated Workflow for Demonstration
This workflow has been simplified to work on other computers.

The watch me do action that tiles the windows has been removed.

I have also renamed the images using exiftool instead of automator as it is unreliable.

I have also changed the file location to be moved to the desktop for replicability